reorganizing utils funcs
