+++ src/renate/utils/file.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 import logging
 import os
-import pickle as pkl
 import shutil
 from pathlib import Path
 from typing import List, Optional, Tuple, Union
@@ -12,8 +11,8 @@ from zipfile import ZipFile
 import boto3
 import pandas as pd
 import requests
-import torch
 from botocore.exceptions import ClientError
+from pytorch_lightning.utilities.rank_zero import rank_zero_only
 
 logger = logging.getLogger(__name__)
 
@@ -281,13 +280,17 @@ def save_pandas_df_to_csv(df: pd.DataFrame, file_path: Union[str, Path]) -> pd.D
     return df
 
 
-def convert_to_tensor(obj):
-    """This function converts a pickleable object to a torch tensor. This is only to
-    aid saving with Deepspeed."""
-    return torch.as_tensor(list(pkl.dumps(obj)))
+@rank_zero_only
+def unlink_file_or_folder(path: Path) -> None:
+    """Funtion to remove files and folders.
 
-
-def recover_object_from_tensor(tensor):
-    """This function converts a tensor to a byte stream that is passed through pickle
-    to recover the underlying object. For usage with Deepspeed"""
-    return pkl.loads(bytes(tensor.tolist()))
+    Unlink works for files, rmdir for empty folders, but not for non-empty ones. Hence a
+    recursive solution.
+    """
+    if path.exists():
+        if path.is_file():
+            path.unlink(missing_ok=True)
+        else:
+            for child in path.iterdir():
+                unlink_file_or_folder(child)
+            path.rmdir()
